Moved `pageSize` to action payload
