+++ ts/features/idpay/initiative/details/screens/OperationsListScreen.tsx
@@ -1,4 +1,3 @@
-import * as pot from "@pagopa/ts-commons/lib/pot";
 import { useRoute } from "@react-navigation/core";
 import { RouteProp } from "@react-navigation/native";
 import { View as NBView } from "native-base";
@@ -14,15 +13,14 @@ import { heightPercentageToDP } from "react-native-responsive-screen";
 import { Body } from "../../../../../components/core/typography/Body";
 import { H1 } from "../../../../../components/core/typography/H1";
 import { IOStyles } from "../../../../../components/core/variables/IOStyles";
+import LoadingSpinnerOverlay from "../../../../../components/LoadingSpinnerOverlay";
 import BaseScreenComponent from "../../../../../components/screens/BaseScreenComponent";
 import I18n from "../../../../../i18n";
-import { useIODispatch, useIOSelector } from "../../../../../store/hooks";
 import customVariables from "../../../../../theme/variables";
 import { formatDateAsLocal } from "../../../../../utils/dates";
 import { TimelineOperationCard } from "../components/TimelineTransactionCards";
 import { IDPayDetailsParamsList } from "../navigation";
-import { idpayTimelineSelector } from "../store";
-import { idpayTimelinePageGet } from "../store/actions";
+import { useInitiativeTimelineFetcher } from "../utils/hooks";
 
 export type OperationsListScreenParams = { initiativeId: string };
 type OperationsListScreenRouteProps = RouteProp<
@@ -55,44 +53,12 @@ const renderLoader = (isLoading: boolean) =>
     />
   ) : null;
 
-const useTimelineFetcher = (initiativeId: string) => {
-  const dispatch = useIODispatch();
-  const timelineFromSelector = useIOSelector(idpayTimelineSelector);
-
-  const isLoading = pot.isLoading(timelineFromSelector);
-  const isError = pot.isError(timelineFromSelector);
-
-  const timeline = pot.getOrElse(timelineFromSelector, {
-    lastUpdate: new Date(),
-    operationList: []
-  });
-  const page = React.useRef(0);
-
-  const fetchNextPage = () => {
-    if (!isError && !isLoading) {
-      page.current = page.current + 1;
-      dispatch(
-        idpayTimelinePageGet.request({ initiativeId, page: page.current })
-      );
-    }
-  };
-
-  return {
-    isLoading,
-    timeline,
-    fetchNextPage
-  } as const;
-};
-
 export const OperationsListScreen = () => {
   const route = useRoute<OperationsListScreenRouteProps>();
   const { initiativeId } = route.params;
 
-  const {
-    isLoading,
-    timeline,
-    fetchNextPage: nextPage
-  } = useTimelineFetcher(initiativeId);
+  const { isLoading, isFirstLoading, timeline, fetchNextPage } =
+    useInitiativeTimelineFetcher(initiativeId, 10);
 
   return (
     <BaseScreenComponent
@@ -101,37 +67,39 @@ export const OperationsListScreen = () => {
       )}
       goBack={true}
     >
-      <SafeAreaView>
-        <View style={IOStyles.horizontalContentPadding}>
-          <H1>
-            {I18n.t(
-              "idpay.initiative.details.initiativeDetailsScreen.configured.operationsList.title"
-            )}
-          </H1>
-          <Body>
-            {I18n.t(
-              "idpay.initiative.details.initiativeDetailsScreen.configured.operationsList.lastUpdated"
-            )}
-            <Body weight="SemiBold">
-              {formatDateAsLocal(timeline.lastUpdate, true)}
+      <LoadingSpinnerOverlay isLoading={isFirstLoading} loadingOpacity={100}>
+        <SafeAreaView>
+          <View style={IOStyles.horizontalContentPadding}>
+            <H1>
+              {I18n.t(
+                "idpay.initiative.details.initiativeDetailsScreen.configured.operationsList.title"
+              )}
+            </H1>
+            <Body>
+              {I18n.t(
+                "idpay.initiative.details.initiativeDetailsScreen.configured.operationsList.lastUpdated"
+              )}
+              <Body weight="SemiBold">
+                {formatDateAsLocal(timeline.lastUpdate, true)}
+              </Body>
             </Body>
-          </Body>
-        </View>
-        <NBView spacer large />
-        <View style={styles.flatListContainer}>
-          <FlatList
-            style={IOStyles.horizontalContentPadding}
-            data={timeline.operationList}
-            keyExtractor={item => item.operationId}
-            renderItem={({ item }) =>
-              TimelineOperationCard({ transaction: item })
-            }
-            onEndReached={nextPage}
-            onEndReachedThreshold={0.5}
-            ListFooterComponent={() => renderLoader(isLoading)}
-          />
-        </View>
-      </SafeAreaView>
+          </View>
+          <NBView spacer large />
+          <View style={styles.flatListContainer}>
+            <FlatList
+              style={IOStyles.horizontalContentPadding}
+              data={timeline.operationList}
+              keyExtractor={item => item.operationId}
+              renderItem={({ item }) =>
+                TimelineOperationCard({ transaction: item })
+              }
+              onEndReached={() => fetchNextPage()}
+              onEndReachedThreshold={0.5}
+              ListFooterComponent={() => renderLoader(isLoading)}
+            />
+          </View>
+        </SafeAreaView>
+      </LoadingSpinnerOverlay>
     </BaseScreenComponent>
   );
 };
